Suspend fn added

+++ app/src/main/java/com/training/coroutinektor/Globalscope.kt
@@ -4,6 +4,7 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.async
 import kotlinx.coroutines.runBlocking
 
+// Run block example
 fun main() = runBlocking {
     val deferred = (1..1_000).map { number ->
         GlobalScope.async {
@@ -11,6 +12,6 @@ fun main() = runBlocking {
         }
     }
 
-    val sum = deferred.map { it.await().toLong() }.sum()
+    val sum = deferred.map { it.await() }.sum()
     print("Sum: $sum")
 }
